+++ src/test/java/nextstep/courses/service/SessionServiceTest.java
@@ -1,14 +1,15 @@
-package nextstep.courses.domain;
+package nextstep.courses.service;
 
+import nextstep.courses.domain.MockSessionRepository;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-public class SessionRegisterTest {
+public class SessionServiceTest {
 
     @Test
     void 강의상태가_모집중일_경우만_저장() {
-        SessionRegister sessionRegister = new SessionRegister(new MockSessionRepository());
+        SessionService sessionRegister = new SessionService(new MockSessionRepository());
 
         sessionRegister.registerSession(1L);
         sessionRegister.registerSession(2L);
